+++ tests/unit/eyes_selenium/test_webdriver.py
@@ -8,7 +8,7 @@ from applitools.selenium.useragent import OSNames
 @pytest.mark.parametrize("version, major, minor", [("9", 9, -1), ("9.1", 9, 1)])
 def test_driver_useragent_splits_version(version, major, minor):
     driver_mock, eyes_mock = MagicMock(), MagicMock()
-    driver_mock.desired_capabilities = {
+    driver_mock.capabilities = {
         "platformName": "Android",
         "deviceName": "Pixel",
         "platformVersion": version,
@@ -22,7 +22,7 @@ def test_driver_useragent_splits_version(version, major, minor):
 
 def test_driver_app_useragent_lowercase_os():
     driver_mock, eyes_mock = MagicMock(), MagicMock()
-    driver_mock.desired_capabilities = {
+    driver_mock.capabilities = {
         "app": "1.zip",
         "platformName": "ios",
         "platformVersion": "13.4",